refactor: perform session gating asynchronously
