+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/comms/delivery/EmbraceDeliveryServiceTest.kt
@@ -4,6 +4,7 @@ import com.google.common.util.concurrent.MoreExecutors
 import io.embrace.android.embracesdk.EmbraceEvent
 import io.embrace.android.embracesdk.comms.api.ApiService
 import io.embrace.android.embracesdk.fakeBackgroundActivity
+import io.embrace.android.embracesdk.fakes.FakeGatingService
 import io.embrace.android.embracesdk.logging.InternalEmbraceLogger
 import io.embrace.android.embracesdk.ndk.NdkService
 import io.embrace.android.embracesdk.payload.Event
@@ -15,6 +16,7 @@ import io.mockk.every
 import io.mockk.mockk
 import io.mockk.verify
 import org.junit.After
+import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
 import org.junit.BeforeClass
 import org.junit.Test
@@ -29,6 +31,7 @@ internal class EmbraceDeliveryServiceTest {
     companion object {
         private lateinit var mockDeliveryCacheManager: EmbraceDeliveryCacheManager
         private lateinit var apiService: ApiService
+        private lateinit var gatingService: FakeGatingService
         private lateinit var logger: InternalEmbraceLogger
 
         @BeforeClass
@@ -38,6 +41,7 @@ internal class EmbraceDeliveryServiceTest {
             every { mockDeliveryCacheManager.loadCrash() } returns null
             every { mockDeliveryCacheManager.getAllCachedSessionIds() } returns emptyList()
             apiService = mockk(relaxed = true)
+            gatingService = FakeGatingService()
             logger = InternalEmbraceLogger()
         }
     }
@@ -46,12 +50,14 @@ internal class EmbraceDeliveryServiceTest {
     fun after() {
         clearAllMocks()
         executor.shutdown()
+        gatingService.sessionMessagesFiltered.clear()
     }
 
     private fun initializeDeliveryService() {
         deliveryService = EmbraceDeliveryService(
             mockDeliveryCacheManager,
             apiService,
+            gatingService,
             executor,
             executor,
             logger
@@ -68,7 +74,8 @@ internal class EmbraceDeliveryServiceTest {
 
         deliveryService.saveSession(mockSessionMessage)
 
-        verify { mockDeliveryCacheManager.saveSession(mockSessionMessage) }
+        verify(exactly = 1) { mockDeliveryCacheManager.saveSession(mockSessionMessage) }
+        assertEquals(1, gatingService.sessionMessagesFiltered.size)
     }
 
     @Test
@@ -164,6 +171,7 @@ internal class EmbraceDeliveryServiceTest {
             )
         }
         verify { mockFuture wasNot Called }
+        assertEquals(1, gatingService.sessionMessagesFiltered.size)
     }
 
     @Test
@@ -188,6 +196,7 @@ internal class EmbraceDeliveryServiceTest {
             )
         }
         verify { mockFuture wasNot Called }
+        assertEquals(1, gatingService.sessionMessagesFiltered.size)
     }
 
     @Test
@@ -214,6 +223,7 @@ internal class EmbraceDeliveryServiceTest {
         verify(exactly = 1) {
             mockFuture.get(1L, TimeUnit.SECONDS)
         }
+        assertEquals(1, gatingService.sessionMessagesFiltered.size)
     }
 
     @Test
